fix: extract pure map helpers out of GlobalMap
GlobalMap.tsx mixed pure computation (building cluster feature
collections, building show/fair lookup maps, nearest-point search,
tab lookup) with component state and effects. Move the pure parts
into helpers/ so GlobalMap only keeps the state wiring, and share the
zoom-level constants via mapZoomLevels.ts so the clustering helper
doesn't need to import from GlobalMap.tsx.
