+++ src/app/Scenes/Map/mapZoomLevels.ts
@@ -1,0 +1,3 @@
+export const DefaultZoomLevel = 11
+export const MinZoomLevel = 9
+export const MaxZoomLevel = 17.5
